requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
awscli==1.32.44
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.35.26
botocore==1.35.26
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.3.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dataclasses-json==0.6.4
dbus-python==1.2.18
dill==0.3.8
discord.py==2.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.15
docutils==0.16
ec2-hibinit-agent==1.0.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
git-remote-codecommit==1.17
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
hibagent==1.0.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.20.3
hyperlink==21.0.0
idna==3.3
ikp3db==1.4.2
importlib-metadata==4.6.4
incremental==21.3.0
isort==5.12.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.3.0
langchain-aws==0.2.1
langchain-community==0.3.0
langchain-core==0.3.5
langchain-text-splitters==0.3.0
langsmith==0.1.127
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
orjson==3.10.7
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.17.4
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.8.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==4.0.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sos==4.5.6
SQLAlchemy==2.0.27
sqlparse==0.4.4
ssh-import-id==5.11
streamlit==1.31.1
systemd-python==234
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.18
validators==0.22.0
wadllib==1.3.6
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0
